requirements.txt
Flask==2.3.3
PyYAML==6.0.1
pytest==7.4.0
gunicorn==21.2.0
python-dotenv==1.0.0
pymodbus==3.5.4
